+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 * ui-orders: module warnings analysis. Refs UIOR-969.
 * Add copy icon to PO and POL number. Refs UIOR-966.
 * Display "Renewal note" field on POL. Refs UIOR-965.
+* Creating an order from instance record. Refs UIOR-973.
+* Cancel creating an order or order line from instance record. Refs UIOR-975.
 
 ## [3.1.2](https://github.com/folio-org/ui-orders/tree/v3.1.2) (2022-03-25)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.1.1...v3.1.2)